requirements.txt
Flask
pandas
openai
chat2plot
langchain-community
plotly
python-dotenv
